FileRouter - code reorganization
