+++ packages/lib/src/router/index.ts
@@ -1,6 +1,6 @@
-export * from "./config.js"
+export { useFileRouter, type FileRouterContextType } from "./context.js"
 export * from "./errors.js"
 export { FileRouter, type FileRouterProps } from "./fileRouter.js"
-export { useFileRouter } from "./context.js"
 export * from "./link.js"
-export * from "./types.js"
+export * from "./pageConfig.js"
+export type * from "./types.js"
