+++ sandbox/csr/src/pages/blog/a/index.tsx
@@ -1,0 +1,7 @@
+export default function BlogPageA() {
+  return (
+    <div>
+      <h1>Blog Page A (root level)</h1>
+    </div>
+  )
+}
